Refactor scale module
