+++ crates/image-ops/src/scale/mod.rs
@@ -1,0 +1,8 @@
+mod filter;
+mod pixel_format;
+#[allow(clippy::module_inception)]
+mod scale;
+
+pub use self::filter::*;
+pub use self::pixel_format::*;
+pub use self::scale::*;
